fixed a size_t to unsigned int conversion warning on 64 bit windows

+++ PLYGeoConverter.cpp
@@ -288,8 +288,8 @@ Mesh* PLYGeoConverter::ConvertToMesh(const std::string& strFilename) {
       }
 
       if (bEdgeColorsFound)  {
-        COLIndices.push_back(colors.size());
-        COLIndices.push_back(colors.size());
+        COLIndices.push_back(UINT32(colors.size()));
+        COLIndices.push_back(UINT32(colors.size()));
         colors.push_back(color);
       }
